learn_python_hard_way/ex51/gothonweb/tests/app_tests.py
# -*- coding: UTF-8 -*-


'''
Last modified time: 2014-12-12 17:35:38
Edit time: 2014-12-12 17:36:03
File name: tools.py
Edit by caimaoy
'''

from nose.tools import *
from bin.app import app
from tests.tools import assert_response

def test_index():
    # check that we get a 404 on the / URL
    resp = app.request("/")
    assert_response(resp, status="404")

    # test our first GET request to /hello
    resp = app.request("/hello")
    assert_response(resp)

    # make sure default values work for the form
    resp = app.request("/hello", method="POST")
    # print resp
    assert_response(resp, contains="Nobody")

    # test that we get expected values
    data = {'name': 'Zed', 'greet': 'Hola'}
    resp = app.request("/hello", method="POST", data=data)
    assert_response(resp, contains="Zed")
